# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  api.Error:
    properties:
      errors:
        additionalProperties: {}
        type: object
    type: object
...
  handler.GenericDataListResponse-oscalTypes_1_1_3_Role:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Role'
        type: array
    type: object
...
  handler.GenericDataResponse-oscalTypes_1_1_3_Role:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscalTypes_1_1_3.Role'
        description: Items from the list response
    type: object
...
  oscalTypes_1_1_3.Link:
    properties:
      href:
        type: string
      media-type:
        type: string
      rel:
        type: string
      resource-fragment:
        type: string
      text:
        type: string
    type: object
...
  oscalTypes_1_1_3.Property:
    properties:
      class:
        type: string
      group:
        type: string
      name:
        type: string
      ns:
        type: string
      remarks:
        type: string
      uuid:
        type: string
      value:
        type: string
    type: object
...
  oscalTypes_1_1_3.Role:
    properties:
      description:
        type: string
      id:
        type: string
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      remarks:
        type: string
      short-name:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Continuous Compliance Framework.
  title: Continuous Compliance Framework API
  version: "1"
paths:
...
  /oscal/roles:
    get:
      description: Retrieves all roles.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: List roles
      tags:
      - Oscal
  /oscal/roles/{id}:
    get:
      description: Retrieves a single Role by its unique ID.
      parameters:
      - description: Party ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Get a Role
      tags:
      - Oscal
...
produces:
- application/json
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/auth/token
    type: oauth2
swagger: "2.0"